Esturcutra cambio de puesto
Todas las posibles combinaciones de cambio de puesto han sido considerada, y a nivel base de datos funcionan, aunque termina redirigiendo a error.jsp sin querer. y como extra, optimicé la carga de la página verUsuarios.jsp

+++ web/verEquipos.jsp
@@ -129,7 +129,7 @@
                                         </a>
                                         </div>
                                         <div class='card-footer text-center'>
-                                            <a class='btn btn-info text-center' href='eliminarEquipo?id=<%=eq.getIDEquipo()%>'>
+                                            <a class='btn btn-info text-center' href='eliminarEquipo?id=<%=eq.getIDEquipo()%>' onclick="return confirmationEquipo()>
                                                 Eliminar equipo
                                             </a>
                                         </div>
